A cleaner and updated version
There was a refactor from @JeffreyWay in this file on main repository before my PR was accepted.

So, I solved the resulting conflicts and developed a cleaner version of the proposed feature, based on his last commit.

Also, I removed the option for ommiting the second argument. While slightly handy for the end user, the implementation was not nice code.